Refactor serial number code to make main less cluttered.
This commit also modifies how the serial number register is accessed
to simplyify use and implementation.
